utils/send_image_via_serial/requirements.txt
Pillow>=9.0.0
pyserial>=3.5
